Added linkage feature to testapi frontend.
Added functions addSubCode() and removeSubCode() to apitest.html
Adding a subcode entails getting all data for the parent code adding a codeID to the subCodeID list or creating a new list with the added item, and then call an update on the parent code.
Also adapted the testfrontend to a renamed API name: qdacity instead of codeendpoint.

Dirty fix for "jumping" navbar depending on scrollbar being active or not. Scrollbar is now always active.

Closes #5