Linked Codesystem UI buttons selected code
When a new code is created while one was selected in the treeview, then the new code becomes a subcode of the selected one.
When clicking the trashcan, the currently selected code is removed.
When the edit button is clicked the ID no longer has to be provided, and is instead loaded from the currently selected item.

Closes #24