Modify and Remove Document Data
The user may now use the document UI buttons to edit the title of a document and remove the document from the project.
The selected document is identified through the jQuery Selectable. Each item now has an additional attribute docID to not confuse index in the selectable with the document ID.