Added coding view
When a code is selected, and the properties button is clicked, a footer is shown containing a datatable of all codings for this code.
For the table we used a jQuery plugin called TataTables.
The codings are found through a query on the html of all documents - more efficient ways may have to be implemented at some point.

Closes #38