Added an event handler for changed state of codesystem UI
When the code system UI changes state, for instance by selecting a different code, a new event handler will be triggered, that sets a global variable for the currently active code. Also, it will be checked if the codings UI is currently visible, and if so, the data table containing the codings for the currently selected code will be updated.

Closes #44