Replaced tinyMCE with Squire
Changed the WYSIWYG text editor to Suire. Compared to tinyMCE this is incredibly lightweight.
Modified Squire to accept coding as an inline tag. Style is passed as an attribute, which highlights codings through the bg color.
Also resized the texteditor using the viewport height, and added an eventlistener to the window resize event to resize the text editor.

Closes #50 #51 #7