Responsiveness of text editor
The text editor responds to the available space whether the coding view is shown at the bottom or not. An offset is calculated in resizeElements() for this. When the codings view is displayed or hidden resizeHandler is called instead of resizeElements directly. This is needed to accommodate the time it takes for the animations to finish.

Closes #52