Added Rudimentary Text Editor Controls
Added controls to format text in the text editor with the options bold, italic, underlined, font and font size. Also added a new save button for the text editor. The input for fontsize and font face are jQuery-UI widgets, the rest are fontawesome buttons. The buttons are not yet responsive to what is selected in the text.

Also added a theme CSS for jQuery widgets

Closes #53