Added toggle for editing document
when the edit toggle in the new project section of the UI is triggered, the div containing the texteditor toolbar is collapsed/shown. The toggle icon is changed via CSS. A function is hooked to the collapse event of the toolbar setting a new readonly property on the text editor.

Closes #40