Added coding for multiple paragraphs
adding a coding to the currently selected text moved to a function of the squire editor. The new implementation ensures that codings spanning multiple nodes are split up into multiple nodes. This has to be taken into account when extracting the codings from the text. multiple tags with the same ID may now exist and will have to be merged to create the full text range of the coding.

Fixes #46