Added tooltips for codings
Added custom tooltips for coding. Tooltipster would not work, because the positioning within the iFrame would be wrong. A simple offset would not work either, since the iFrame may contain scrollable conent.

Also refactored getActiveID() to getActiveCode() returning an object with id and name as attributes.
The name attribute of a code is now also set as an attribute of each coding, and will be read when the tooltips are set.

Closes #73