Added properties view
The codings view is now split in two tabs one for the codings as before, and one for the code properties showing the code name and the author.
Currently the properties can not be edited this way, but are simply displayer.
Added the dependency on easytabs because the jQuery UI tabs impose additional CSS which is then automatically used in the generated HTML which I don't want.

Closes #29