Added 401 Error Page
When one of the API responds with a 401 error code during the launch of the coding editor, he will be redirected to a 401 error page instead of showing the uninitialized coding editor.
The 401 code is currently only checked explicitly when loading the project, not for each API call. Each call however is protected in any case without this explicit handling, this just adds a visual representation for users using our application as intended, and not trying to hack the system.

Closes #85