User Invitation
On the project dashboard a user may now invite another user that is already in the DB to join the project. The invited user is identified by his email address.
On the personal dashboard there is now a notification widget where new invitations to projects appear. If the user confirms, then the project will be added to his list of projects.

Closes #105