Added Code Memos
There now is a new tab in the codings view, which displays a code memo in a squire instance. This instance has to be created with the onload event of the iFrame. This impedes performance but according to documentation this is neccessary. Adjusted squire-raw.js to work in firefox when instantiation squire in the hidden tab.

Closes #81