Added Coding Brackets
Content of the coding editor now leaves a margin on the left side. This space is now used by a SVG drawing brackets around the code elements within the text. Brackets are drawn using D3. Brackets are only set when the document is loaded.
Also moved the main CSS of the editor's iframe to its own file.

Closes #119