Visualized Overlapping Coding Brackets
Whenever a new coding bracket is created it is checked if it overlaps any other coding bracket. If so, then the vertical line of the coding bracket is moved to the left. The label is also checked for other labels it may be overlapping. The label position is moved further down until enough free space is detected.

Closes #121