Linked Coding Bracket path and label
when the user hovers over a coding bracket label, the corresponding path will be highlighted. For this both are now under an new svg element that groups them. Then when the mouseover event is triggered, the hover class is assigned and the element is reordered to be at the bottom of the root SVG element. This is due to SVG ignoring z-index values, and instead relies purely on the order within the DOM (at least in some browsers). The reordering ensures that the highlighted bracket is drawn on top.

Closes #122