Frontend Build Pipeline
Set up a transpiling pipeline. JS code is piped through Babel with Gulp using the ES2015 preset. This resolves ES6 specific code like classes, promises, export/import. But the export/import modules are transpiled to require() statements which are still useless in the browser. Therefore the code is then bundled using webpack into a single entry point for each page of the app. The whole pipeline uses a bunch of npm modules, which are documented in the package.json and may be installed using npm install --dev. Async loading of gapi is handled using scriptjs.

Closes #166