Intercoder Agreement
Implemented F-Measure calculation on paragraph level between a ValidationProject and the corresponding revision. Added a new API method for this, which can be called from the project dashboard.
Changed the id that is put into the marked up text to an ID that is unique only relative to the project. This way, when the project is copied to a new revision snapshot the codesystem can simply be copied, meaning the codes all receive new IDs in the database, but they retain their codeID attribute which is referenced in the text, so the text does not have to be updated.
Changed logging level to INFO.

Closes #190