Project Dashboard for ValidationProjects
The project dashboard now supports displaying validationprojects.
Added new API method listValidationCoders, which lists all validationcoders for a validationproject in the users panel on the project dashboard.
Both types of projects (PROJECT/VALIDATION) now have one conditional panel, that is by default collapsed and only shown when the appropriate type parameter is recognized in the URL.
Fixed a few type parameters in links.

Closes #202