Updated List of Validationprojects
The Link to the coding editor of the validation project is now only added to elements with the class validationProjectLink. This is only the case if the user is Admin, the owner of the original project or validationcoder in the specific validation project.

the delete button is only available for owners of the original project or admins.

Closes #248, Closes #247