Added Validation Reports
the results of the intercoder validation are now stored in a new database entity called ValidationReport which in turn holds references to a set of ValidationResults, which hold the values for the f-measure.

Added an element listing the validation reports to the revision history. New reports can be created through a button in the section for validationprojects. A table containing all the different validationresults can be accessed through a modal opened when the report is clicked in the list of reports.

Closes #252