Prevented re-initialization, fixed codebook editor
When the session takes long enough without any interaction so the user is re-authenticated due to the listener on the google account object the code system view and the document view is not re-initialized if already existing.

Fixed a typo causing the code book editor to not being initialized correctly.

Fixes #258, Fixes #256