Added loading image to landing page
On the landing page there will now be a loading animation shown as soon as the user hits the "sign in with google" button. The loading animation is a react component which is loaded in a hidden div on loading the page. If it is mounted only when the button is clicked it takes 2 seconds to show, that is why it is loaded in the hidden div.

Closes #259